requirements.txt
# Core Flask and web framework
Flask>=3.1.0
Flask-SocketIO>=5.3.6
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.6.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-Mail>=0.9.1

# Database and authentication
supabase>=2.3.4
python-dotenv>=1.0.0

# HTTP and networking
requests>=2.32.3
httpx>=0.25.2
httpcore>=1.0.8
urllib3>=2.4.0
certifi>=2025.1.31

# Data processing and analysis - using more compatible versions
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
openai>=1.75.0

# PDF processing
PyPDF2>=3.0.1
reportlab>=4.0.4
pdfkit>=1.0.0

# Image processing
Pillow>=11.2.1

# Utilities
click>=8.1.8
tqdm>=4.67.1
markdown2>=2.5.3
python-dateutil>=2.9.0.post0

# Payment processing
stripe>=12.0.0

# Email services
mailjet-rest>=1.3.4

# WebSocket and async
eventlet>=0.35.1
websockets>=12.0

# Data validation and serialization
pydantic>=2.11.3
pydantic-core>=2.33.1

# Development and deployment
gunicorn>=23.0.0

# Additional dependencies
anyio>=4.9.0
blinker>=1.9.0
Brotli>=1.1.0
cffi>=1.17.1
charset-normalizer>=3.4.1
contourpy>=1.3.2
cssselect2>=0.8.0
cycler>=0.12.1
distro>=1.9.0
fonttools>=4.57.0
h11>=0.14.0
idna>=3.10
itsdangerous>=2.2.0
Jinja2>=3.1.6
jiter>=0.9.0
kiwisolver>=1.4.8
MarkupSafe>=3.0.2
packaging>=25.0
pycparser>=2.22
pydyf>=0.11.0
pyparsing>=3.2.3
pyphen>=0.17.2
six>=1.17.0
sniffio>=1.3.1
tinycss2>=1.4.0
tinyhtml5>=2.0.0
typing-inspection>=0.4.0
typing_extensions>=4.13.2
webencodings>=0.5.1
Werkzeug>=3.1.3
zopfli>=0.2.3.post1

# Celery and Redis
celery>=5.2
redis>=4.0
